Split apart graphics/mod.rs
